Refactor and add logging
