--- 06-crud/README.md
@@ -1,5 +1,0 @@
-# 6. CRUD
-
-- Add Create & Get endpoints
-- Remove seed data
-
