+++ 06-fatal/README.md
@@ -1,0 +1,5 @@
+# 6. Fatal
+
+- Create `run() error` function
+- Remove all calls to `log.Fatal`
+
