+++ 06-fatal/seed-data.sql
@@ -1,0 +1,3 @@
+INSERT INTO products (product_id, name, quantity, cost)
+VALUES	('a2b0639f-2cc6-44b8-b97b-15d69dbb511e', 'Comic Books', 50, 42),
+	('72f8b983-3eb4-48db-9ed0-e45cc6bd716b', 'McDonalds Toys', 75, 120);
