--- 07-business-logic-tests/README.md
@@ -1,5 +1,0 @@
-# 7. Business-Logic Tests
-
-- Add test incorporating Create, Get, & List
-- Add test helper function for setting up the database
-
