+++ 07-logging/README.md
@@ -1,0 +1,5 @@
+# 7. Logging
+
+- Create `internal/platform/log` package
+- Replace all calls to stdlib `log` package
+
