+++ 07-logging/schema.sql
