+++ 08-crud/README.md
@@ -1,0 +1,7 @@
+# 6. CRUD
+
+- Add `Create` & `Get` endpoints
+- Add `ServeHTTP` method with router
+- Update `Handler` in `main`
+- Remove seed data
+
