+++ 08-crud/schema.sql
@@ -1,0 +1,8 @@
+CREATE TABLE products (
+	product_id	UUID,
+	name		VARCHAR(255),
+	cost		INT,
+	quantity	INT,
+
+	PRIMARY KEY (product_id)
+);
