+++ 09-refactor-handlers/README.md
@@ -1,0 +1,5 @@
+# 7. Refactor Handlers
+
+- Add `internal/platform/web` package
+- Update handler methods to match `web.HandlerFunc`
+
