+++ 09-refactor-handlers/internal/products/products.go
