+++ 10-business-logic-tests/README.md
@@ -1,0 +1,5 @@
+# 7. Business-Logic Tests
+
+- Add `internal/products` test incorporating Create, Get, & List
+- Add test helper function for setting up the database
+
